Improve student usability on MakeEntry screen
Before, it wasn't clear which pay period a student was creating an
entry for. Now the student can see the status of all the entry status of
all the recent pay periods and select which one they want to create new
entries for.

This commit also fixes the issue where s "Save Success" message would be displayed
before the save actually succeeded. Indeed, if the save failed, the student
would still see the success message. Now the form is disabled while saving is
in progress, and the success message is only shown on actual success.